+++ packages/esl/src/all.ts
@@ -63,3 +63,6 @@ export * from './esl-lazy-template/core';
 
 // Line Clamp
 export * from './esl-line-clamp/core';
+
+// JSX Custom Tags Types Declarations
+export * from './types/jsx.shape';
