+++ packages/esl/src/esl-a11y-group/core/esl-a11y-group.shape.ts
@@ -15,10 +15,3 @@ export interface ESLA11yGroupTagShape extends ESLBaseElementShape<ESLA11yGroup> 
   /** Children are not allowed for ESLA11yGroup */
   children?: never[];
 }
-
-declare global {
-    export interface ESLIntrinsicElements {
-      /** {@link ESLA11yGroup} custom tag */
-      'esl-a11y-group': ESLA11yGroupTagShape;
-    }
-}
