+++ packages/esl/src/esl-anchornav/core.ts
@@ -1,3 +1,5 @@
+export type {ESLAnchornavTagShape} from './core/esl-anchornav.shape';
+
 export * from './core/esl-anchornav';
 export * from './core/esl-anchornav-sticked';
 export * from './core/esl-anchor';
