+++ packages/esl/src/esl-base-element/core/esl-base-element.shape.ts
@@ -5,11 +5,3 @@ import type {AriaAttributesShape} from '../shape/wai-aria.shape';
 /** Defines the shape of the Base Custom Element for TSX (strongly typed JSX) */
 export interface ESLBaseElementShape<T> extends HTMLElementShape, DOMEventsAttributesShape<T>, AriaAttributesShape {
 }
-
-declare global {
-  export interface ESLIntrinsicElements {}
-
-  namespace JSX {
-    interface IntrinsicElements extends ESLIntrinsicElements {}
-  }
-}
