+++ packages/esl/src/esl-footnotes/core/esl-footnotes.shape.ts
@@ -18,10 +18,3 @@ export interface ESLFootnotesTagShape extends ESLBaseElementShape<ESLFootnotes> 
   /** Children are not allowed for ESLShare */
   children?: never[];
 }
-
-declare global {
-  export interface ESLIntrinsicElements {
-    /** {@link ESLFootnotes} custom tag */
-    'esl-footnotes': ESLFootnotesTagShape;
-  }
-}
