+++ packages/esl/src/esl-media/core/esl-media.shape.ts
@@ -62,10 +62,3 @@ export type ESLMediaTagShape = ({
   /** Children are not allowed for ESLMedia */
   children?: never[];
 } & ESLBaseElementShape<ESLMedia>;
-
-declare global {
-  export interface ESLIntrinsicElements {
-    /** {@link ESLMedia} custom tag */
-    'esl-media': ESLMediaTagShape;
-  }
-}
