+++ packages/esl/src/esl-tooltip/core/esl-tooltip.shape.ts
@@ -6,10 +6,3 @@ import type {ESLTooltip} from './esl-tooltip';
  * Used for TSX declaration
  */
 export interface ESLTooltipTagShape<T extends ESLTooltip = ESLTooltip> extends ESLPopupTagShape<T> {}
-
-declare global {
-  export interface ESLIntrinsicElements {
-    /** {@link ESLTooltip} custom tag */
-    'esl-tooltip': ESLTooltipTagShape;
-  }
-}
